+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## [5.0.0-beta.12](https://github.com/maevsi/sqitch/compare/5.0.0-beta.11...5.0.0-beta.12) (2025-02-16)
+
+### ⚠ BREAKING CHANGES
+
+* **role:** read usernames from secrets (#49)
+
+### Features
+
+* **role:** read usernames from secrets ([#49](https://github.com/maevsi/sqitch/issues/49)) ([2a413f6](https://github.com/maevsi/sqitch/commit/2a413f62150859a6f90aabe1f1ee03dc6ddf0382)), closes [#154](https://github.com/maevsi/sqitch/issues/154)
+
+## [5.0.0-beta.11](https://github.com/maevsi/sqitch/compare/5.0.0-beta.10...5.0.0-beta.11) (2025-02-06)
+
+### ⚠ BREAKING CHANGES
+
+* **event:** allow to mark upload as header image (#144)
+
+### Features
+
+* **event:** allow to mark upload as header image ([#144](https://github.com/maevsi/sqitch/issues/144)) ([25e76dd](https://github.com/maevsi/sqitch/commit/25e76ddb4256ef5d0fa6b15cdc16dba00a79bf17))
+
 ## [5.0.0-beta.10](https://github.com/maevsi/sqitch/compare/5.0.0-beta.9...5.0.0-beta.10) (2025-02-04)
 
 ### Bug Fixes